release: 发布 v5.0.0（64 位时间戳 flag-day + 统一迁移工具链）
major bump：tstamp / expiry_at 全链路 uint32_t → uint64_t（Y2038 readiness）。
ABI / 盘上格式破坏——SONAME libbitcask.so.4 → .so.5，旧二进制由链接器层隔离；
bitcask.meta v3 → v4 门禁干净拒开 u32 纪元旧库（v1/v2/v3），但不必重建——
新增 bitcask_migrate tstamp64 子命令提供 u32 → u64 非破坏性离线迁移。

单一真源 CMakeLists.txt project(VERSION) 4.1.0 → 5.0.0，派生处同步：
bitcask_version.h（C API bitcask_version_*）、libbitcask.so VERSION/SOVERSION、
CHANGELOG header、README 版本段。

主要变更（详见 CHANGELOG.md）：

- tstamp / expiry_at 全链路 u64：C/C++ API 签名 + 盘上 record header（23B→27B）
  + DocValue v3→v4 + hint v3→v4（magic BCH3→BCH4）+ keydir 快照 v2→v3
  + docmap sidecar v2→v3 + kDocmapDeltaV2→V3 + kCkptVersion2→3
  + segment_v2 v1→v2（DocRow 字段重排）+ DocStore v1→v2
- 新增 bitcask_migrate 统一迁移入口（detect / be2le / tstamp64 三个子命令）
  + 库 API bitcask::migrate::migrate_u32_to_u64()——u32 纪元库非破坏性迁移
  + DocValue v3→v4 转码（Ver 字节 + expiry 段 u32→u64 零扩展）
  + meta v2/v3 → v4（除 version 字节 + CRC 外逐字节照搬）
- Fixed：u32 wrap 致全库 key 误判过期（expiry_secs 极大时求和回绕）
  → get / iter / merge_policy cutoff 三处统一 u64 域算术
- 回归测试 ×2：HugeExpirySecsNoWrap + Post2106TimestampRoundTrip
- v1/v2 旧版本读端从「兼容读」翻为「干净拒收」（meta v4 是统一门禁）
- 测试 golden 字节换代（data record / DocValue / hint v4）
- ASan 构建 -Wshorten-64-to-32 修复：data_file_test 局部 Rec/R 结构体 ts 字段同步 u64

验收：ASan Debug 全目标构建绿 + 相关测试套件 290/290 通过。

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ c_api/bitcask_kv.cpp
@@ -191,7 +191,7 @@ BITCASK_API bitcask_error_t bitcask_get(bitcask_t* cask,
 BITCASK_API bitcask_error_t bitcask_put(bitcask_t* cask,
                                           bitcask_slice_t key,
                                           bitcask_slice_t value,
-                                          uint32_t tstamp,
+                                          uint64_t tstamp,
                                           bitcask_fault_t* fault) {
     // S13-M2：extern "C" 异常隔离
     return guarded(fault, [&]() -> bitcask_error_t {
@@ -213,8 +213,8 @@ BITCASK_API bitcask_error_t bitcask_put(bitcask_t* cask,
 BITCASK_API bitcask_error_t bitcask_put_ex(bitcask_t* cask,
                                            bitcask_slice_t key,
                                            bitcask_slice_t value,
-                                           uint32_t tstamp,
-                                           uint32_t expiry_at,
+                                           uint64_t tstamp,
+                                           uint64_t expiry_at,
                                            bitcask_fault_t* fault) {
     // S13-M2：extern "C" 异常隔离
     return guarded(fault, [&]() -> bitcask_error_t {
@@ -233,7 +233,7 @@ BITCASK_API bitcask_error_t bitcask_put_ex(bitcask_t* cask,
 
 BITCASK_API bitcask_error_t bitcask_delete(bitcask_t* cask,
                                              bitcask_slice_t key,
-                                             uint32_t tstamp,
+                                             uint64_t tstamp,
                                              bitcask_fault_t* fault) {
     // S13-M2：extern "C" 异常隔离
     return guarded(fault, [&]() -> bitcask_error_t {
@@ -289,7 +289,7 @@ BITCASK_API void bitcask_get_result_free(bitcask_get_result_t* result) {
 BITCASK_API bitcask_error_t bitcask_put_doc(bitcask_t* cask,
                                               bitcask_slice_t key,
                                               const bitcask_doc_input_t* doc,
-                                              uint32_t tstamp,
+                                              uint64_t tstamp,
                                               bitcask_fault_t* fault) {
     // S13-M2：extern "C" 异常隔离
     return guarded(fault, [&]() -> bitcask_error_t {
@@ -328,7 +328,7 @@ BITCASK_API void bitcask_search_result_batch_free(bitcask_search_result_t** resu
 BITCASK_API bitcask_error_t bitcask_put_batch(bitcask_t* cask,
                                               const bitcask_kv_pair_t* items,
                                               size_t n,
-                                              uint32_t tstamp,
+                                              uint64_t tstamp,
                                               bitcask_fault_t* fault) {
     // S13-M2：extern "C" 异常隔离
     return guarded(fault, [&]() -> bitcask_error_t {
